create sale working

+++ sales/api/sales_rest/models.py
@@ -21,7 +21,7 @@ class Customer(models.Model):
 
 
 class Sale(models.Model):
-    price = models.PositiveSmallIntegerField()
+    price = models.PositiveIntegerField()
     automobile = models.ForeignKey(
         AutomobileVO,
         related_name= 'sale',
